pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.isort]
profile = "black"

[tool.poetry]
authors = ["liblaf <i@liblaf.me>"]
description = ""
license = "MIT"
name = "taichi-extras"
packages = [{ include = "taichi_extras" }]
readme = "README.md"
repository = "https://github.com/liblaf/taichi-extras"
version = "0.1.0"

[tool.poetry.dependencies]
matplotlib = "^3.7.2"
meshio = { extras = ["all"], version = "^5.3.4" }
meshtaichi-patcher = "^0.0.19"
open3d = "^0.17.0"
opencv-python = "^4.8.0.76"
pydantic = "^2.3.0"
pydicom = "^2.4.3"
pymeshfix = "^0.16.2"
pynrrd = "^1.0.0"
python = ">=3.10,<3.11"
pyvista = { extras = ["all"], version = "^0.42.1" }
requests = "^2.31.0"
taichi = "^1.6.0"
trimesh = { extras = ["all"], version = "^3.23.5" }
typer = { extras = ["all"], version = "^0.9.0" }

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
